Remove `import_deep_distortion_by_type()`` and replace with new `write_distorted_inputs()`, add tests etc.
By replacing `import_deep_distortion()` with this, we now just rely on the `excluded_charges` in the `low_energy_defects` dict to identify the other charge states for which the low-energy distorted structures should be tested, which contains the previously _tested_ charge states (in `get_deep_distortions()`) for which this distortion wasn't found (which also means if a certain charge state wasn't parsed correctly then it won't be included (as if it were, this could add unnecessary duplicate calculations for cases where it _was_ found for that charge state, but just wasn't parsed right)(i.e. only include charge states for which we've actually tested with structure matching)).
Also, it means we don't require the `ChargedDefectsStructures()`-generated `defect_dict` object as input, which likely wouldn't be loaded in the notebook at this point, because we're running this after doing the initial calculations, and also makes it readily implemented as a CLI function. (In one go can identify energy-lowering distortions and write the corresponding test calc input files).


Former-commit-id: c0c68dbff00a3fa43f8b5b7d443334e0b0bdcc7e